+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+ - Tuple counterparts of variadics (`tuple.h`):
+   - `ML99_tupleCount` (`ML99_TUPLE_COUNT`).
+   - `ML99_tupleIsSingle` (`ML99_TUPLE_IS_SINGLE`).
+   - `ML99_tupleForEach(I)`.
+
 ## [1.5.0] - 2021-08-11
 
 ### Added
 
- - `ML99_assignInitializerList(Stmt)` as `ML99_assign(Stmt)` counterparts for initializer lists.
+ - `ML99_assignInitializerList(Stmt)` as `ML99_assign(Stmt)` counterparts for initialiser lists.
 
 ## [1.4.1] - 2021-08-05